# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Curriculum Activity & Attendance App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timetable:
    get:
      operationId: listTimetable
      tags: [timetable]
      summary: List timetable entries
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
        - name: dayOfWeek
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      operationId: createTimetableEntry
      tags: [timetable]
      summary: Create a timetable entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimetableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
  /timetable/{id}:
    patch:
      operationId: updateTimetableEntry
      tags: [timetable]
      summary: Update a timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimetableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
    delete:
      operationId: deleteTimetableEntry
      tags: [timetable]
      summary: Delete a timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /timetable/today:
    get:
      operationId: getTodayTimetable
      tags: [timetable]
      summary: Get today's timetable for the current user's class
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Today's timetable
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"

  # CURRICULUM (SUBJECTS)
...
components:
  schemas:
...
    TimetableEntry:
      type: object
      required: [id, subjectId, facultyId, department, year, section, dayOfWeek, startTime, endTime]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        subjectCode:
          type: ["string", "null"]
        facultyId:
          type: ["integer", "null"]
        facultyName:
          type: ["string", "null"]
        department:
          type: string
        year:
          type: integer
        section:
          type: string
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
        isCurrent:
          type: boolean
    TimetableInput:
      type: object
      required: [subjectId, department, year, section, dayOfWeek, startTime, endTime]
      properties:
        subjectId:
          type: integer
        facultyId:
          type: integer
        department:
          type: string
        year:
          type: integer
        section:
          type: string
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
    TimetableUpdate:
      type: object
      properties:
        subjectId:
          type: integer
        facultyId:
          type: integer
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
...